requirements.txt
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
torchmetrics
librosa
numpy==1.26.4
speechbrain
tqdm
matplotlib
pandas
wandb
comet_ml
transformers
hydra-core

black
isort
pre-commit
